+++ transport/src/test/java/io/netty/channel/DefaultChannelPipelineTest.java
@@ -331,7 +331,7 @@ public class DefaultChannelPipelineTest {
         pipeline.addBefore("1", "0", newHandler());
         pipeline.addAfter("10", "11", newHandler());
 
-        AbstractChannelHandlerContext ctx = (AbstractChannelHandlerContext) pipeline.firstContext();
+        DefaultChannelHandlerContext ctx = (DefaultChannelHandlerContext) pipeline.firstContext();
         assertNotNull(ctx);
         while (ctx != null) {
             int i = toInt(ctx.name());
@@ -1583,8 +1583,8 @@ public class DefaultChannelPipelineTest {
         }
     }
 
-    private static int next(AbstractChannelHandlerContext ctx) {
-        AbstractChannelHandlerContext next = ctx.next;
+    private static int next(DefaultChannelHandlerContext ctx) {
+        DefaultChannelHandlerContext next = ctx.next;
         if (next == null) {
             return Integer.MAX_VALUE;
         }
@@ -1607,11 +1607,16 @@ public class DefaultChannelPipelineTest {
         pipeline.executor().submit(new Runnable() {
             @Override
             public void run() {
-                AbstractChannelHandlerContext ctx = (AbstractChannelHandlerContext) pipeline.firstContext();
+                DefaultChannelHandlerContext ctx = (DefaultChannelHandlerContext) pipeline.firstContext();
                 int handlerNumber = 0;
-                while (ctx != ((DefaultChannelPipeline) pipeline).tail) {
-                    handlerNumber++;
-                    ctx = ctx.next;
+                if (ctx != null) {
+                    for (;;) {
+                        handlerNumber++;
+                        if (ctx == pipeline.lastContext()) {
+                            break;
+                        }
+                        ctx = ctx.next;
+                    }
                 }
                 assertEquals(expectedNumber, handlerNumber);
             }
